inventory/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from knox import views as knox_views
from .views import RegisterAPI, LoginAPI, ProductViewSet

router = DefaultRouter()
router.register(r'products', ProductViewSet, basename='product')

urlpatterns = [
    path('', include(router.urls)),
    path('auth/register/', RegisterAPI.as_view(), name='register'),
    path('auth/login/', LoginAPI.as_view(), name='login'),
    path('auth/logout/', knox_views.LogoutView.as_view(), name='logout'),
]