requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
